Migrate 'fillable' test
